filtered search added to items at path /search . It can take params search_text and tags via a form submit (which are prioritized) and the same keys as body keys for searches we do not want reloading the page.
item.js has also had it's check for missing checks buffed and reinstated along with checking if price is a number.

as seen with the banner and table that prints out when search settings are made to be submitted to prisma, I made a console log wrapper that you can use to wrap your text in inside a console log if you need something to be ultra visibile. It has red,blue,green,black, and magenta backgrounds hardcoded in, though it defaults to black if nothing is specified. Use them wisely.